programasSena/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cors==1.0.1
distlib==0.3.7
Django==4.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.12.2
future==0.18.3
gevent==23.7.0
greenlet==2.0.2
idna==3.4
platformdirs==3.10.0
pycparser==2.21
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==3.4.4
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.3
zope.event==5.0
zope.interface==6.0
